refactor: move commands into cmd directory
