+++ go.mod
@@ -17,7 +17,6 @@ require (
 	github.com/cosmos/cosmos-sdk v0.50.13
 	github.com/hashicorp/go-metrics v0.5.4
 	github.com/spf13/cobra v1.9.1
-	github.com/spf13/viper v1.20.0
 	github.com/stretchr/testify v1.10.0
 	github.com/tendermint/tendermint v0.0.0-00010101000000-000000000000
 	golang.org/x/sync v0.12.0
@@ -164,6 +163,7 @@ require (
 	github.com/spf13/afero v1.12.0 // indirect
 	github.com/spf13/cast v1.7.1 // indirect
 	github.com/spf13/pflag v1.0.6 // indirect
+	github.com/spf13/viper v1.20.0 // indirect
 	github.com/subosito/gotenv v1.6.0 // indirect
 	github.com/syndtr/goleveldb v1.0.1-0.20220721030215-126854af5e6d // indirect
 	github.com/tendermint/go-amino v0.16.0 // indirect
